% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harrietr.R
\docType{package}
\name{harrietr}
\alias{harrietr}
\alias{harrietr-package}
\title{\code{harrietr} package}
\description{
harrietr: Wrangle Phylogenetic Distance Matrices and Other Utilities
}
\details{
See the README on
\href{https://cran.r-project.org/package=harrietr/README.html}{CRAN}
or \href{https://github.com/andersgs/harrietr#readme}{GitHub}
}
