% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{woodmouse}
\alias{woodmouse}
\title{Woodmouse dataset}
\format{An object of class \code{DNAbin} with 15 rows and 965 columns.}
\source{
"ape" package \link[ape]{woodmouse}
}
\usage{
woodmouse
}
\description{
Woodmouse dataset
}
\keyword{datasets}

