% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{woodmouse_meta}
\alias{woodmouse_meta}
\title{Woodmouse metadata}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15 rows and 2 columns.}
\usage{
woodmouse_meta
}
\description{
A dummy metadata table generated to demonstrate the use of \code{add_metadata}.
}
\keyword{datasets}

